\name{rat}
\docType{data}
\alias{rat}
\title{Blood pressure measures of the rats}
\description{
The description of each column of the data set is as follows:

id2 = the id variable for each rat that is provided in Davis (2002).

id = a new id variable that takes value 1,2,..43 after sorting on id and group.

time = the timing of each measurement.

group = the group variable that takes value 1, 2, 3, or 4

bp = the blood pressure value.

group1 = indicator variables for group 1 that takes value one for rats in group 1 and that takes value 0 otherwise.

group2 = indicator variables for group 2 that takes value one for
rats in group 2 and that takes value 0 otherwise.

group3 = indicator variables for group 3 that takes value one for rats in group 3 and that takes value 0 otherwise.

group4 = indicator variables for group 4 that takes value one for rats in group 4 and that takes value 0 otherwise.

highbp = a variable that takes value 1 if the rat's blood pressure is at least 100.
}
\usage{rat}
\format{A data frame contains 291 observations and 10 variables.}
\source{Table 6.11 of Davis (2002)}
\references{
Davis, C.(2002). \emph{Statistical Methods for the Analysis of Repeated Measurements.}
}
\keyword{datasets}